fix: #WB-2215, refactor and typings of ResourceService; removed explorer interface
