+++ src/ts/globals.ts
@@ -18,71 +18,36 @@ export const ERROR_CODE = {
 } as const;
 export type ErrorCode = (typeof ERROR_CODE)[keyof typeof ERROR_CODE];
 
-//-- Applications
+//-- Core applications (but many others exist).
 export const APP = {
-  EXPLORER: "explorer",
+  ADMIN: "admin",
+  ARCHIVE: "archive",
+  AUTH: "auth",
+  CAS:"cas",
+  COMMUNICATION: "communication",
+  CONVERSATION: "conversation",
+  DIRECTORY: "directory", // FIXME userbook OR directory : the choice may impact some configurations, @see IXitiTrackingParams.NOM_PAGE for example
+  USERBOOK: "userbook",
+  INFRA: "infra",
   PORTAL: "portal",
-  BLOG: "blog",
-  EXERCIZER: "exercizer",
   TIMELINE: "timeline",
-  CAS: "cas",
-  VIDEO: "video",
-  MINDMAP: "mindmap",
-
-  // TODO compléter/trier les apps suivantes
-  /*
-  "competences"
-  "cahier-textes"
-  "poll"
-  "rack"
-  "rbs"
-  "searchengine"
-  "sharebigfiles"
-  "schoolbook"
-  "archive"
-  "admin"
-  "cahier-de-texte"
-  "wiki"
-  "cns"
-  "conversation"
-  "paths"
-  "parcours"
-  "notebook"
-  "scrap-book"
-  "scrapbook"
-  "account"
-  "support"
-  "workspace"
-  "admin-portal"
-  "stats"
-  "userbook"    // FIXME userbook OR directory : the choice may impact some configurations, @see IXitiTrackingParams.NOM_PAGE for example
-  "directory"   // FIXME Keep in mind that ode-ts-client MUST not access the locationPath of the window !
-  "mindmap"
-  "timelinegenerator"
-  "actualites"
-  "pad"
-  "collaborativeeditor"
-  "settings-class"
-  "library"
-  "visioconf"
-  "Web-conference"
-  "notes"
-  "attendance"
-  "calendar"
-  "canal-numerique"
-  "collaborative-wall"
-  "statistics"
-  "polls"
-  "community"
-  "forum"
-  "pages"
-  "website"
-  "parametrage"
-  "kne"
-  "sacoche"
-*/
+  WORKSPACE: "workspace",
+  // -- a few others commonly used apps
+  EXPLORER: "explorer",
 } as const;
-export type App = (typeof APP)[keyof typeof APP]; // type App = "explorer" | "blog" | "exercizer"...
+export type App = (typeof APP)[keyof typeof APP]; // type App = "admin" | "archive" | "auth"...
+
+//-- Semantical typings
+/**
+ * An ID is a unique string in its applicable field.
+ */
+export type ID = string;
+
+/**
+ * Type of resource generated by an application.
+ * For example "folder", "blog", "mindmap"...
+ */
+export type ResourceType = string;
 
 //-- User preferences keys
 export const USER_PREFS = {
