+++ src/ts/workspace/interface.ts
@@ -1,5 +1,3 @@
-export type ID = string;
-
 export interface WorkspaceElement {
   _id?: string;
   eType: "folder" | "file" | string;
